funciona correctamente el abm de perfil y de usuario

+++ src/main/java/com/springboot/clienteapp/models/repository/UsuarioRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.springboot.clienteapp.models.entity.Usuario;
 
 @Repository
-public interface UsuarioRepository extends CrudRepository<Usuario, Integer> {
+public interface UsuarioRepository extends CrudRepository<Usuario, Long> {
 
 }
